+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * 【json   】     修复JSONGetter.getJSONArray判断null的问题（issue#I4C15H@Gitee）
 * 【db     】     修复Condition没占位符的情况下sql没引号问题（issue#1846@Github）
 * 【cache  】     修复FIFOCache中remove回调无效问题（pr#1856@Github）
+* 【json   】     修复JSONArray.set中，index为0报错问题（issue#1858@Github）
 
 -------------------------------------------------------------------------------------------------------------